Get rid of json-parser submodule and just check out the correct version of the library manually in the Makefile
The problem with the submodule approach is that the source tarball (both the one built by Github and the one we build ourselved) does not contain the main .git directory.  Because of that, git submodule will refuse to run and the build will fail when trying to use the official source.
Fixes #16