Fine-tune the deb builds (#48)
Compose the debs inside the container (rather than only building inside
the container). This fixes issues with newer deb versions as built on
ubuntu-latest not working on debian 10.
Also fix the location of the binary inside the deb and adjust the
package name for better consistency with aother pam packages.