Change LOG_ERR to LOG_DEBUG for 'No more lines...'
This is not an error condition, because feof(fp) will only be true once one
tries to read *past* the last byte, and not yet after reading the last byte.

Therefore the 'No more lines' condition is expected to be reached every time
a configuration file is read; it is thus no more than a debug message.